🎨 Migrated label filter to use MultiSelectCombobox from Shade
The inline label filter used a custom InlinePopover/InlineList
implementation that duplicated logic already available in Shade's
MultiSelectCombobox. This replaces it with the shared component,
extracts EditRow to its own file, and adds onSearchChange support
to MultiSelectCombobox for external search sync.
