+++ apps/posts/src/components/label-picker/index.ts
@@ -1,3 +1,4 @@
 export {default as LabelPicker} from './label-picker';
 export {default as LabelFilterRenderer} from './label-filter-renderer';
+export {EditRow} from './edit-row';
 export type {LabelPickerProps} from './label-picker';
